requirements.txt
asgiref==3.9.2
dataclasses==0.6
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djongo==1.2.31
dnspython==2.8.0
gunicorn==23.0.0
packaging==25.0
PyJWT==2.10.1
pymongo==4.15.1
setuptools==65.5.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.11.0
